refactor: extract bufferToBlobUrl + encodeWav to audio-fixtures
